Delete multiple records
I ran in to a problem when trying to get password resetting to work. The problem was that when trying to delete the existing password_reset records, the delete() function was called on a query without a recordId, resulting in an incorrect URL being sent to the Filemaker API.
The solution I found was to check for this when calling the delete function and delete the records one at a time if this was the case.
This might not be the safest not the most efficient solution, but at least it solved one of the issues trying to manage password resets with Laravel and Filemaker. I also don't know how else to solve this when the Filemaker API only supports deleting based on recordId.
Any help, comments or feedback on this issue would be greatly appreciated.